kubernetes/backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
pyzmq